feat(save): major simplification of save functionality
