+++ next-tavla/src/Admin/scenarios/Edit/components/BoardTitle/index.tsx
@@ -2,7 +2,6 @@ import { TextField } from '@entur/form'
 import { EditIcon } from '@entur/icons'
 import { ChangeEventHandler, useCallback } from 'react'
 import { useEditSettingsDispatch } from '../../utils/contexts'
-import { Heading2, Heading5 } from '@entur/typography'
 
 function BoardTitle({ title }: { title?: string }) {
     const dispatch = useEditSettingsDispatch()
@@ -21,18 +20,16 @@ function BoardTitle({ title }: { title?: string }) {
     )
 
     return (
-        <div className="flexColumn">
-            <Heading5 as={Heading2}>Navn på tavla</Heading5>
-            <TextField
-                value={title ?? ''}
-                label="Navn på tavla"
-                placeholder="Navn på tavla"
-                aria-label="Endre navn på tavla"
-                prepend={<EditIcon aria-hidden="true" />}
-                onChange={dispatchTitle}
-                ref={autoSelect}
-            />
-        </div>
+        <TextField
+            value={title ?? ''}
+            className="w-30"
+            label="Navn på tavla"
+            placeholder="Navn på tavla"
+            aria-label="Endre navn på tavla"
+            prepend={<EditIcon aria-hidden="true" />}
+            onChange={dispatchTitle}
+            ref={autoSelect}
+        />
     )
 }
